refactor: more sensible file structure
