+++ src/components/events/directory/filter/panels/Location.tsx
@@ -5,7 +5,7 @@ import {
   EventFiltersSchema,
   eventLocationFilterEnum,
 } from '@src/utils/eventFilter';
-import FilterList, { FilterListItem } from '../FilterList';
+import FilterList, { FilterListItem } from '../base/FilterList';
 import { FilterPanelProps, panelProps, setEventsParams } from '../utils';
 
 export type LocationPanelFields = Pick<
